do_prediction*: refactor into separate module
